refactor: 도메인에 모델 패키지 추가 #253
