+++ backend/src/main/java/com/dubu/backend/member/api/response/MemberInfoResponse.java
@@ -1,7 +1,7 @@
 package com.dubu.backend.member.api.response;
 
-import com.dubu.backend.member.domain.Address;
-import com.dubu.backend.member.domain.Member;
+import com.dubu.backend.member.domain.model.Address;
+import com.dubu.backend.member.domain.model.Member;
 import com.dubu.backend.member.domain.enums.AddressType;
 import com.dubu.backend.todo.domain.Category;
