+++ backend/src/main/java/com/dubu/backend/member/api/response/MemberSavedAddressResponse.java
@@ -1,6 +1,6 @@
 package com.dubu.backend.member.api.response;
 
-import com.dubu.backend.member.domain.Address;
+import com.dubu.backend.member.domain.model.Address;
 import com.dubu.backend.member.domain.enums.AddressType;
 
 import java.util.List;
