+++ backend/src/main/java/com/dubu/backend/member/application/MemberQueryFacade.java
@@ -4,8 +4,8 @@ import com.dubu.backend.member.api.response.MemberInfoResponse;
 import com.dubu.backend.member.api.response.MemberSavedAddressResponse;
 import com.dubu.backend.member.api.response.MemberStatusResponse;
 import com.dubu.backend.member.core.exception.MemberSavedAddressNotFoundException;
-import com.dubu.backend.member.domain.Address;
-import com.dubu.backend.member.domain.Member;
+import com.dubu.backend.member.domain.model.Address;
+import com.dubu.backend.member.domain.model.Member;
 import com.dubu.backend.member.domain.repository.AddressRepository;
 import com.dubu.backend.member.domain.repository.MemberCategoryRepository;
 import com.dubu.backend.member.domain.repository.MemberRepository;
