+++ backend/src/main/java/com/dubu/backend/member/application/MemberServiceHelper.java
@@ -1,6 +1,6 @@
 package com.dubu.backend.member.application;
 
-import com.dubu.backend.member.domain.Member;
+import com.dubu.backend.member.domain.model.Member;
 import com.dubu.backend.member.domain.repository.MemberRepository;
 import com.dubu.backend.member.core.exception.MemberNotFoundException;
