+++ backend/src/main/java/com/dubu/backend/member/domain/repository/MemberCategoryRepository.java
@@ -1,7 +1,7 @@
 package com.dubu.backend.member.domain.repository;
 
-import com.dubu.backend.member.domain.Member;
-import com.dubu.backend.member.domain.MemberCategory;
+import com.dubu.backend.member.domain.model.Member;
+import com.dubu.backend.member.domain.model.MemberCategory;
 import com.dubu.backend.todo.domain.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
