+++ backend/src/main/java/com/dubu/backend/member/domain/repository/MemberLocationRepository.java
@@ -1,6 +1,6 @@
 package com.dubu.backend.member.domain.repository;
 
-import com.dubu.backend.member.domain.MemberLocation;
+import com.dubu.backend.member.domain.model.MemberLocation;
 
 public interface MemberLocationRepository {
     void saveMemberLocation(Long memberId, MemberLocation memberLocation);
