+++ backend/src/main/java/com/dubu/backend/member/domain/repository/MemberRepository.java
@@ -1,6 +1,6 @@
 package com.dubu.backend.member.domain.repository;
 
-import com.dubu.backend.member.domain.Member;
+import com.dubu.backend.member.domain.model.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
