+++ backend/src/main/java/com/dubu/backend/plan/domain/Plan.java
@@ -1,7 +1,7 @@
 package com.dubu.backend.plan.domain;
 
 import com.dubu.backend.core.domain.BaseTimeEntity;
-import com.dubu.backend.member.domain.Member;
+import com.dubu.backend.member.domain.model.Member;
 import com.dubu.backend.plan.core.AggregateRoot;
 import jakarta.persistence.*;
 import lombok.*;
