+++ backend/src/main/java/com/dubu/backend/plan/domain/repository/SubPathRepository.java
@@ -1,6 +1,6 @@
 package com.dubu.backend.plan.domain.repository;
 
-import com.dubu.backend.member.domain.Member;
+import com.dubu.backend.member.domain.model.Member;
 import com.dubu.backend.plan.domain.SubPath;
 import com.dubu.backend.plan.domain.Plan;
 import com.dubu.backend.todo.domain.enums.TodoType;
