+++ backend/src/main/java/com/dubu/backend/todo/application/impl/save/SaveTodoManagementService.java
@@ -1,6 +1,6 @@
 package com.dubu.backend.todo.application.impl.save;
 
-import com.dubu.backend.member.domain.Member;
+import com.dubu.backend.member.domain.model.Member;
 import com.dubu.backend.member.domain.enums.Status;
 import com.dubu.backend.member.core.exception.MemberNotFoundException;
 import com.dubu.backend.member.domain.repository.MemberRepository;
