+++ backend/src/main/java/com/dubu/backend/todo/application/impl/statistic/StatisticServiceImpl.java
@@ -1,6 +1,6 @@
 package com.dubu.backend.todo.application.impl.statistic;
 
-import com.dubu.backend.member.domain.Member;
+import com.dubu.backend.member.domain.model.Member;
 import com.dubu.backend.member.core.exception.MemberNotFoundException;
 import com.dubu.backend.member.domain.repository.MemberRepository;
 import com.dubu.backend.plan.domain.Feedback;
