+++ backend/src/main/java/com/dubu/backend/todo/domain/Schedule.java
@@ -1,7 +1,7 @@
 package com.dubu.backend.todo.domain;
 
 import com.dubu.backend.core.domain.BaseTimeEntity;
-import com.dubu.backend.member.domain.Member;
+import com.dubu.backend.member.domain.model.Member;
 import jakarta.persistence.*;
 import lombok.*;
