+++ backend/src/main/java/com/dubu/backend/todo/dto/request/TodoCreateRequest.java
@@ -1,6 +1,6 @@
 package com.dubu.backend.todo.dto.request;
 
-import com.dubu.backend.member.domain.Member;
+import com.dubu.backend.member.domain.model.Member;
 import com.dubu.backend.plan.domain.SubPath;
 import com.dubu.backend.todo.domain.*;
 import com.dubu.backend.todo.domain.enums.TodoDifficulty;
