+++ backend/src/main/java/com/dubu/backend/todo/infra/repository/querydsl/CustomTodoRepository.java
@@ -1,7 +1,5 @@
 package com.dubu.backend.todo.infra.repository.querydsl;
 
-import com.dubu.backend.member.domain.Member;
-import com.dubu.backend.todo.dto.response.MemberCategoryInfo;
 import com.dubu.backend.todo.dto.common.Cursor;
 import com.dubu.backend.todo.dto.search.TodoSearchCond;
 import com.dubu.backend.todo.domain.Todo;
@@ -9,7 +7,6 @@ import com.dubu.backend.todo.domain.enums.TodoType;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 
-import java.time.LocalDate;
 import java.util.List;
 
 public interface CustomTodoRepository {
