+++ backend/src/test/java/com/dubu/backend/member/unit/application/MemberCommandFacadeTest.java
@@ -4,12 +4,12 @@ import com.dubu.backend.member.application.MemberCommandFacade;
 import com.dubu.backend.member.application.MemberLocationFacade;
 import com.dubu.backend.member.application.MemberQueryFacade;
 import com.dubu.backend.member.application.event.MovementCompletedEvent;
-import com.dubu.backend.member.domain.Address;
-import com.dubu.backend.member.domain.Member;
-import com.dubu.backend.member.domain.MemberCategory;
+import com.dubu.backend.member.domain.model.Address;
+import com.dubu.backend.member.domain.model.Member;
+import com.dubu.backend.member.domain.model.MemberCategory;
 import com.dubu.backend.member.domain.enums.AddressType;
 import com.dubu.backend.member.domain.enums.Status;
-import com.dubu.backend.member.domain.MemberLocation;
+import com.dubu.backend.member.domain.model.MemberLocation;
 import com.dubu.backend.member.api.request.MemberInfoUpdateRequest;
 import com.dubu.backend.member.api.request.MemberOnboardingRequest;
 import com.dubu.backend.member.api.response.MemberInfoResponse;
