+++ backend/src/test/java/com/dubu/backend/plan/application/SubPathFacadeTest.java
@@ -1,6 +1,6 @@
 package com.dubu.backend.plan.application;
 
-import com.dubu.backend.member.domain.Member;
+import com.dubu.backend.member.domain.model.Member;
 import com.dubu.backend.member.core.exception.MemberNotFoundException;
 import com.dubu.backend.member.domain.repository.MemberRepository;
 import com.dubu.backend.plan.domain.Path;
