+++ backend/src/test/java/com/dubu/backend/todo/application/impl/path/SubPathTodoQueryServiceTest.java
@@ -8,7 +8,7 @@ import java.util.List;
 import java.util.Optional;
 
 import com.dubu.backend.core.domain.PageResponse;
-import com.dubu.backend.member.domain.Member;
+import com.dubu.backend.member.domain.model.Member;
 import com.dubu.backend.member.domain.enums.Status;
 import com.dubu.backend.member.core.exception.MemberNotFoundException;
 import com.dubu.backend.member.domain.repository.MemberCategoryRepository;
