+++ backend/src/test/java/com/dubu/backend/todo/application/impl/statistic/StatisticServiceImplTest.java
@@ -12,7 +12,7 @@ import java.util.List;
 import java.util.Optional;
 
 import com.dubu.backend.member.domain.enums.OauthProvider;
-import com.dubu.backend.member.domain.Member;
+import com.dubu.backend.member.domain.model.Member;
 import com.dubu.backend.member.domain.enums.Role;
 import com.dubu.backend.member.domain.enums.Status;
 import com.dubu.backend.member.core.exception.MemberNotFoundException;
