+++ backend/src/test/java/com/dubu/backend/todo/application/impl/today/TodayTodoManagementServiceTest.java
@@ -1,7 +1,7 @@
 package com.dubu.backend.todo.application.impl.today;
 
 import com.dubu.backend.member.domain.enums.OauthProvider;
-import com.dubu.backend.member.domain.Member;
+import com.dubu.backend.member.domain.model.Member;
 import com.dubu.backend.member.domain.enums.Role;
 import com.dubu.backend.member.domain.enums.Status;
 import com.dubu.backend.member.core.exception.MemberNotFoundException;
