+++ backend/src/test/java/com/dubu/backend/todo/application/impl/today/TodayTodoQueryServiceTest.java
@@ -1,6 +1,5 @@
 package com.dubu.backend.todo.application.impl.today;
 
-import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
 import static org.junit.jupiter.api.Assertions.*;
 
@@ -9,7 +8,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
-import com.dubu.backend.member.domain.Member;
+import com.dubu.backend.member.domain.model.Member;
 import com.dubu.backend.member.domain.enums.Status;
 import com.dubu.backend.member.domain.repository.MemberRepository;
 import com.dubu.backend.member.domain.repository.MemberCategoryRepository;
