+++ backend/src/test/java/com/dubu/backend/todo/application/impl/tomorrow/TomorrowTodoManagementServiceTest.java
@@ -11,7 +11,7 @@ import java.util.List;
 import java.util.Optional;
 import java.util.ArrayList;
 
-import com.dubu.backend.member.domain.Member;
+import com.dubu.backend.member.domain.model.Member;
 import com.dubu.backend.member.domain.enums.Status;
 import com.dubu.backend.member.domain.enums.Role;
 import com.dubu.backend.member.domain.enums.OauthProvider;
